requirements.txt
# Core dependencies
pyaedt>=0.7.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0

# Data handling and conversion
h5py>=3.6.0
scikit-rf>=0.24.0

# Configuration and environment
python-dotenv>=0.19.0
pyyaml>=6.0

# Logging and utilities
loguru>=0.6.0
tqdm>=4.62.0

# Data visualization
plotly>=5.3.0
seaborn>=0.11.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
pylint>=2.12.0

# Documentation
sphinx>=4.3.0
sphinx-rtd-theme>=1.0.0

# Optional: For advanced features
# jupyter>=1.0.0
# notebook>=6.4.0
